Adding support for load balancers
If your Django instance is behind some proxy, load balancer, etc. the  REMOTE_ADDR would be the IP of the proxy.  In that case HTTP_X_FORWARDED_FOR is set with either the REMOTE_ADDR from the client, or if HTTP_X_FORWARDED_FOR already exists it appends a comma and the REMOTE_ADDR of the proxy that it received the request from.  The Client's IP address should always be first if there is an HTTP_X_FORWARDED_FOR header.

As an extreme example, say you are using an LB, and the client is behind a proxy (maybe they work at a bank or big corp).  The headers will look like this:
"REMOTE_ADDR": "10.6.34.123" (your LB IP)
"HTTP_X_FORWARDED_FOR": "10.12.250.12,54.65.23.162" (the client's internal IP, the public IP of the proxy)

What I did was have it check for that HTTP_X_FORWARDED_FOR header, grab the first ip address in the list, and if it doesn't exist it will fall back to the REMOTE_ADDR header.  You can still override that with RESTRICTEDSESSIONS_REMOTE_ADDR_KEY if you want to.

Also added some debug messages for troubleshooting if people are getting dupmed unexpectedly